tests/wpt/tests/html/browsers/history/the-history-interface/combination_history_001.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>combination_history_001(Combine pushState and replaceSate methods.)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    test(function () {
      window.history.pushState(1, document.title, '?x=1');
      assert_equals(history.state, 1, "first");

      window.history.replaceState(2, document.title, '?x=1');
      assert_equals(history.state, 2, "second")
    }, "Combine pushState and replaceSate methods");
    </script>
  </body>
</html>
